+++ app/react/V2/Routes/Settings/IX/helpers/specs/helpers.spec.ts
@@ -3,8 +3,8 @@
  */
 
 import * as translate from 'app/I18N/translateFunction';
-import { formatOptions } from '../../ExtractorModal';
-import { formatExtractors } from '../../../IXDashboard';
+import { formatOptions } from '../../components/ExtractorModal';
+import { formatExtractors } from '../../IXDashboard';
 import { getAvailableSources } from '../helpers';
 import { extractors, templates, templatesWithCommonProperties } from './fixtures';
