+++ app/react/V2/Routes/Settings/IX/helpers/specs/updateSuggestionsByEntity.spec.ts
@@ -1,4 +1,4 @@
-import { MultiValueSuggestion } from '../../../types';
+import { MultiValueSuggestion } from '../../types';
 import { updateSuggestionsByEntity } from '../helpers';
 import {
   suggestion1,
